pyproject.toml
[project]
name = "alb-priority-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.241.0",
    "constructs>=10.5.1",
]
